Simplify the workflow app layout
