+++ examples/apps/durable_agent_workflows/.gitignore
@@ -1,0 +1,1 @@
+.workflow-data/
