--- examples/apps/durable_agent_workflows/backend/workflow_agent/__init__.py
@@ -1,5 +1,0 @@
-import vercel.workflow
-
-# The app uses one registry for all workflow decorators so queue messages
-# are dispatched by the same Workflows instance.
-workflow = vercel.workflow.Workflows()
